feat(payment):  Get User Payment Status

+++ backend/src/schemas/payment.py
@@ -1,10 +1,11 @@
+from decimal import Decimal
 from pydantic import BaseModel, Field
 from typing import Literal
 from src.models.payment import PaymentGatewayEnum
 
 
 class PaymentCreate(BaseModel):
-    amount: int
+    amount: Decimal
     shipping_address_id: int
     gateway: Literal["mock", "razorpay"] = Field(default="mock")
     simulate_success: bool | None = None
@@ -13,7 +14,7 @@ class PaymentCreate(BaseModel):
 class PaymentOut(BaseModel):
     id: int
     order_id: int
-    amount: int
+    amount: Decimal
     status: str
     is_paid: bool
     payment_gateway: PaymentGatewayEnum
